# Remove for loop in parseSize to enable inlining